3.1.7 zmiana nazwy pakietu, 3.1.8 przeniesienie wspólnej zmiennej klas IntArrayStack oraz IntLinkedList do interfejsu
