+++ src/main/java/edu/kis/vh/nursery/stack/Interface.java
@@ -1,6 +1,8 @@
-package edu.kis.vh.nursery.list;
+package edu.kis.vh.nursery.stack;
 
 public interface Interface {
+
+    int STACK_VALUE = -1;
 
     void push(int i);
